requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
google-cloud-firestore
google-cloud-pubsub
